+++ Source/HtmlRenderer.PdfSharp/Adapters/GraphicsPathAdapter.cs
@@ -36,7 +36,7 @@ namespace HtmlRenderer.PdfSharp.Adapters
         /// <summary>
         /// Appends an elliptical arc to the current figure.
         /// </summary>
-        public void AddArc(float x, float y, float width, float height, float startAngle, float sweepAngle)
+        public void AddArc(double x, double y, double width, double height, double startAngle, double sweepAngle)
         {
             _graphicsPath.AddArc(x, y, width, height, startAngle, sweepAngle);
         }
@@ -44,7 +44,7 @@ namespace HtmlRenderer.PdfSharp.Adapters
         /// <summary>
         /// Appends a line segment to this GraphicsPath.
         /// </summary>
-        public void AddLine(float x1, float y1, float x2, float y2)
+        public void AddLine(double x1, double y1, double x2, double y2)
         {
             _graphicsPath.AddLine(x1, y1, x2, y2);
         }
